third_party/blink/web_tests/http/tests/devtools/sources/debugger-ui/show-generator-location-expected.txt

Tests that "Show Generator Location" jumps to the correct location.


Running: testiterNotStarted
Generator location revealed: [17:20]

Running: testiterSuspended1
Generator location revealed: [19:11]

Running: testiterSuspended2
Generator location revealed: [20:11]

Running: testiterSuspended3
Generator location revealed: [21:11]

Running: testiterClosed
Generator location revealed: [17:20]
